backend/requirements.txt
fastapi
langchain
langchain-core
langchain-community
langchain-ollama
python-dotenv
requests
beautifulsoup4
uvicorn
langchain-google-genai